sheet 3

+++ Sheet#3/A_Summation.c
@@ -3,7 +3,7 @@ int main(){
     int N;
     scanf("%d",&N);
     int A[N];
-    int sum =0;
+    long long sum =0;
     for (int i = 0; i < N; i++)
     {
         scanf("%d",&A[i]);
@@ -14,7 +14,7 @@ int main(){
         sum = -(sum);
     }
     
-    printf("%d",sum);
+    printf("%lld",sum);
     
     
     return 0;
